* lisp/ox-linguistics.el: remove half-start on handling for other item   commands
I've decided that what I've got is enough.  The user can already
override the default labeling scheme by providing a list tag; and
supporting the other commands doesn't gain us much, since these
commands seem to be pretty broken.

For example, gb4e's \exp command breaks label/ref.  Although it does
produce a "primed" example in the output PDF, it doesn't give you a
way to refer to that primed example; you can't say

\exp{foo}{\label{foo-prime}My primed example}

and then later \ref{foo-prime}.  The reference doesn't appear in the
output. (Argh!)

Producing the correct output for the primed example in this case can
already be done using the list tag mechanism (which exports to the
\exi command in gb4e), and so nothing is gained by providing support
for \exp as well.

The only case where it would actually be useful to support a different
underlying command is glosses, and I think that can be done in a
different way.